enqueue the notification email to the movie's creator
email processing over SMTP can take several seconds, which is several
seconds a unicorn process would be blocked for. this means your user is
waiting for an additional several seconds for a response, which is poor
user experience.

on a technical note, moving long-running processes out of the request
path is crucial for ensuring a stable site, as its incredibly easy to
starve a unicorn pool if you're waiting on long-running resources in the
middle of a request - new requests will queue on the master, thus
further impacting end user request time.